Workaround for CUDA 11.4.4 build in Conda on Ubuntu 22 / v6 kernel (#3459)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3459

When building with CUDA 11.4.4, CMake does not properly include files under Conda environment. This workaround flattens the include sub-directories in to the include root. It will unblock us for now while we are looking for a fix through CMakeLists files or figure out if it's a CMake bug and it gets fixed.

Reviewed By: junjieqi

Differential Revision: D57545169

fbshipit-source-id: 9cbdd0866e00e899cc889930a59448da55d873c2