Disable the NEON optimisations on gcc <= 8 (#3869)
Summary:
GCC7 doesnt support all the necessary NEON intrinsics, which is really a shame. However this means that for aarch64 GCC cannot compile faiss with neon intrinsics, so we should avoid using them.

This is similar to https://github.com/facebookresearch/faiss/issues/3860, build issues on GCC7, which I need. This one is a bit uglier, since GCC7 does support NEON just not all of the intrinsics.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3869

Reviewed By: asadoughi

Differential Revision: D63081962

Pulled By: junjieqi

fbshipit-source-id: 69827cd447dd405b3ef70d651996f9ad00b6213e